requirements.txt
aws_requests_auth
awscli>=1.42.14
bedrock-agentcore-starter-toolkit==0.1.10
bedrock-agentcore==0.1.3
boto3==1.40.30
botocore==1.40.30
dill>=0.4.0,<0.5.0
faiss-cpu
fpdf
iprogress
ipykernel
ipywidgets
matplotlib
nbformat
opensearch-py
papermill
pillow
pypdf
PyPDF2
requests
retrying
scipy
seaborn
strands-agents-tools==0.2.7
strands-agents==1.8.0
streamlit
termcolor
uv
cfn-flip